feat(settings): carry the cloud badge onto the API key marks
Every provider that can hold a key is one whose sessions live in a cloud
service — that is what makes the key necessary — so the same mark that now
carries a cloud in the session list carries one here too, rather than reading
differently in the two places it appears.

The badge moves next to the mark it rides and each place sizes it against
that mark: a credential line shows a bare mark at text height rather than the
session list's 32px tile, so the badge comes down with it, and the viewBox
carries the cloud's proportions so only a width is ever restated.
