+++ apps/desktop/src/renderer/styles/base.css
@@ -543,6 +543,30 @@ html[data-keyboard="true"] .compact-hover-target:focus-visible {
   transform: scale(0.94);
 }
 
+/* The work is happening somewhere the user cannot see. The badge overhangs the
+   mark's box rather than sitting inside it, so it reads as attached to the mark
+   instead of crowding it, and it is out of flow so the mark stays where it was
+   whether or not a badge is drawn. It fills against the surface rather than
+   against whatever it is laid on: a session row tinted for attention would
+   otherwise recolour the badge, and where the work runs is not a state. Each
+   place that shows a mark sets the size, because a 32px session tile and a
+   text-height settings line cannot carry the same one. */
+.cloud-badge {
+  position: absolute;
+  display: grid;
+  border-radius: 50%;
+  background: var(--surface-fill);
+  box-shadow: inset 0 0 0 1px var(--hairline);
+  color: var(--text-secondary);
+  place-items: center;
+}
+
+/* The viewBox already carries the cloud's proportions, so a place that shows a
+   badge sets a width and the height follows from the shape. */
+.cloud-badge svg {
+  height: auto;
+}
+
 /* Luke's face ------------------------------------------------------------ */
 
 /* The place nearest the housing is the one the capsule keeps, and Luke holds it
@@ -1200,10 +1224,33 @@ html[data-keyboard="true"] .compact-hover-target:focus-visible {
 
 /* Beside the provider's name, sized to it: the mark reads as part of the word,
    not as an icon in a column of its own. */
+/* A bare mark at text height rather than the session list's tile, so the badge
+   comes down with it: any larger and the annotation would outweigh the mark it
+   annotates, and this is the smallest the cloud still reads at. It clears the
+   name beside it because the line's gap is wider than the overhang. */
 .credential-mark {
+  position: relative;
+  display: grid;
   width: 13px;
   height: 13px;
   flex: 0 0 auto;
+  place-items: center;
+}
+
+.credential-mark .provider-mark {
+  width: 13px;
+  height: 13px;
+}
+
+.credential-mark .cloud-badge {
+  right: -3px;
+  bottom: -3px;
+  width: 10px;
+  height: 10px;
+}
+
+.credential-mark .cloud-badge svg {
+  width: 6.5px;
 }
 
 /* The hint sits opposite the editor's buttons and yields to them. */
