+++ apps/desktop/src/renderer/styles/sessions.css
@@ -72,29 +72,17 @@
   height: 17px;
 }
 
-/* The session runs somewhere the user cannot see. The badge overhangs the tile
-   rather than sitting inside it, so it reads as attached to the mark instead of
-   crowding it, and it is taken out of flow so the mark stays centred in the
-   tile whether or not a session carries one. It fills against the surface, not
-   the row: the attention row is tinted, and a badge that changed colour with
-   the row would read as state rather than as where the work runs. */
-.row-cloud {
-  position: absolute;
+/* Against a 32px tile the badge can be read at a glance, so it is sized to be
+   read rather than to be minimal. */
+.row-avatar .cloud-badge {
   right: -4px;
   bottom: -4px;
-  display: grid;
   width: 14px;
   height: 14px;
-  border-radius: 50%;
-  background: var(--surface-fill);
-  box-shadow: inset 0 0 0 1px var(--hairline);
-  color: var(--text-secondary);
-  place-items: center;
 }
 
-.row-cloud svg {
+.row-avatar .cloud-badge svg {
   width: 9px;
-  height: 6.3px;
 }
 
 .row-copy {
